moving tests to BasicTests, adding a test on PropertyFetcherTest.cs and updating changelog
